Move SessionRepository classes to repository package
Removes it from session-service dependency jar, so that application context
probles are avoided.
